+++ src/core.ts
@@ -29,8 +29,6 @@ export {
 } from './rx-collection';
 
 export {
-    _handleFromStorageInstance,
-    _handleToStorageInstance,
     fillObjectDataBeforeInsert
 } from './rx-collection-helper';