+++ src/hooks.ts
@@ -62,7 +62,7 @@ export const HOOKS: { [k: string]: any[] } = {
 
     /**
      * Runs after the document data is ready from
-     * the storage instance.
+     * the RxStorage instance.
      */
     postReadFromInstance: [],