+++ src/plugins/replication-couchdb.ts
@@ -45,7 +45,7 @@ import type {
     SyncOptions,
     PouchDBInstance
 } from '../types';
-import { _handleFromStorageInstance } from '../rx-collection-helper';
+import { transformDocumentDataFromRxStorageToRxDB } from '../rx-storage-helper';
 
 // add pouchdb-replication-plugin
 addPouchPlugin(PouchReplicationPlugin);
@@ -173,7 +173,7 @@ export function setPouchEventEmitter(
 
                 (ev as any).change.docs
                     .filter((doc: any) => doc.language !== 'query') // remove internal docs
-                    .map((doc: any) => _handleFromStorageInstance(rxRepState.collection, doc)) // do primary-swap and keycompression
+                    .map((doc: any) => transformDocumentDataFromRxStorageToRxDB(rxRepState.collection, doc)) // do primary-swap and keycompression
                     .forEach((doc: any) => rxRepState._subjects.docs.next(doc));
             }));